tweak pagination & carousel.
