+++ components/pagination/PaginationPropTypes.tsx
@@ -3,7 +3,6 @@ import * as React from 'react';
 interface PaginationPropTypes {
   prefixCls?: string;
   className?: string;
-  size?: 'large' | 'small';
   mode?: 'button' | 'number' | 'pointer';
   simple?: Boolean;
   style?: React.CSSProperties;
